voice/requirements.txt
gtts>=2.5.0
pyttsx3>=2.90
pydub>=0.25.1
openai>=1.0.0
azure-cognitiveservices-speech>=1.32.0

